Deduplicate lifecycle methods
