+++ pkg/controller/knativeeventingkafkachannel/knativeeventingkafkachannel_controller.go
@@ -128,7 +128,7 @@ func (r *ReconcileKnativeEventingKafkaChannel) Reconcile(request reconcile.Reque
 // Initialize status conditions
 func (r *ReconcileKnativeEventingKafkaChannel) initStatus(instance *operatorv1alpha1.KnativeEventingKafkaChannel) error {
 	if len(instance.Status.Conditions) == 0 {
-		instance.Status.InitializeConditions()
+		operatorv1alpha1.InitializeConditions(&instance.Status)
 		if err := r.updateStatus(instance); err != nil {
 			return err
 		}
@@ -161,14 +161,14 @@ func (r *ReconcileKnativeEventingKafkaChannel) install(instance *operatorv1alpha
 	}
 	r.config.Transform(fns...)
 
-	if instance.Status.IsDeploying() {
+	if operatorv1alpha1.IsDeploying(&instance.Status) {
 		return nil
 	}
 	defer r.updateStatus(instance)
 
 	// Apply the resources in the YAML file
 	if err := r.config.ApplyAll(); err != nil {
-		instance.Status.MarkInstallFailed(err.Error())
+		operatorv1alpha1.MarkInstallFailed(&instance.Status, err.Error())
 		return err
 	}
 
@@ -178,7 +178,7 @@ func (r *ReconcileKnativeEventingKafkaChannel) install(instance *operatorv1alpha
 
 	// Update status
 	instance.Status.Version = version.Version
-	instance.Status.MarkInstallSucceeded()
+	operatorv1alpha1.MarkInstallSucceeded(&instance.Status)
 	log.Info("Install succeeded", "version", version.Version)
 	return nil
 }
@@ -200,19 +200,19 @@ func (r *ReconcileKnativeEventingKafkaChannel) checkDeployments(instance *operat
 		if u.GetKind() == "Deployment" {
 			key := client.ObjectKey{Namespace: u.GetNamespace(), Name: u.GetName()}
 			if err := r.client.Get(context.TODO(), key, deployment); err != nil {
-				instance.Status.MarkDeploymentsNotReady()
+				operatorv1alpha1.MarkDeploymentsNotReady(&instance.Status)
 				if errors.IsNotFound(err) {
 					return nil
 				}
 				return err
 			}
 			if !available(deployment) {
-				instance.Status.MarkDeploymentsNotReady()
+				operatorv1alpha1.MarkDeploymentsNotReady(&instance.Status)
 				return nil
 			}
 		}
 	}
-	instance.Status.MarkDeploymentsAvailable()
+	operatorv1alpha1.MarkDeploymentsAvailable(&instance.Status)
 	log.Info("All deployments are available")
 	return nil
 }
